# ARROW-10588: [Rust] Safe and parallel bit operations for Arrow