--- a/rust/arrow/src/array/array_list.rs
+++ b/rust/arrow/src/array/array_list.rs
@@ -305,7 +305,7 @@ mod tests {
     };
 
     use super::*;
-    use crate::util::bit_ops::BufferBitSliceMut;
+    use crate::util::bit_ops::{BufferBitSlice, BufferBitSliceMut};
 
     #[test]
     fn test_list_array() {
@@ -722,7 +722,7 @@ mod tests {
         assert_eq!(1, sliced_array.offset());
         assert_eq!(2, sliced_array.null_count());
 
-        let null_bit_slice = BufferBitSliceMut::new(&mut null_bits);
+        let null_bit_slice = BufferBitSlice::new(&null_bits);
         for i in 0..sliced_array.len() {
             if null_bit_slice.get_bit(sliced_array.offset() + i) {
                 assert!(sliced_array.is_valid(i));