--- a/rust/arrow/src/array/array_string.rs
+++ b/rust/arrow/src/array/array_string.rs
@@ -25,7 +25,7 @@ use super::{
     Array, ArrayData, ArrayDataRef, GenericListArray, GenericStringIter, LargeListArray,
     ListArray, OffsetSizeTrait,
 };
-use crate::util::bit_util;
+
 use crate::{buffer::Buffer, datatypes::ToByteSlice};
 use crate::{buffer::MutableBuffer, datatypes::DataType};
 
@@ -168,8 +168,7 @@ where
             if let Some(s) = s {
                 let s = s.as_ref();
                 // set null bit
-                let null_slice = null_buf.data_mut();
-                bit_util::set_bit(null_slice, i);
+                null_buf.set_bit(i);
 
                 length_so_far = length_so_far + OffsetSize::from_usize(s.len()).unwrap();
                 offsets.push(length_so_far);