* **Up to %95 improvements on many kernels**.
* Implements safe bit operations for Arrow
* Implements `typed_bits` to get bits as Vec<bool>
* Implements various bit operations for the use with arrow arrays
* Adjusts parquet array reader to use Arrow bit operations

I squashed all my work into a single commit and closing PR: #8598 with comment: https://github.com/apache/arrow/pull/8598#issuecomment-727294219

<details>
<summary>Benchmarks</summary>
<br>
<p>

```
sum 2^20                time:   [160.84 us 161.83 us 163.10 us]                     
                        change: [-88.535% -88.450% -88.365%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe

Benchmarking min 2^20: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.1s, enable flat sampling, or reduce sample count to 50.
min 2^20                time:   [1.4026 ms 1.4027 ms 1.4028 ms]                      
                        change: [-0.0516% -0.0286% -0.0098%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low severe
  8 (8.00%) high mild
  3 (3.00%) high severe

sum nulls 2^20          time:   [783.74 us 797.75 us 812.02 us]                           
                        change: [-80.849% -80.392% -79.903%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild

min nulls 2^20          time:   [12.366 ms 12.375 ms 12.389 ms]                           
                        change: [+24.048% +24.202% +24.381%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe

min string 2^20         time:   [9.9124 ms 9.9145 ms 9.9174 ms]                            
                        change: [+1.9874% +2.0847% +2.1694%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 22 outliers among 100 measurements (22.00%)
  2 (2.00%) low severe
  5 (5.00%) low mild
  8 (8.00%) high mild
  7 (7.00%) high severe

min nulls string 2^20   time:   [19.348 ms 19.352 ms 19.357 ms]                                  
                        change: [+3.4758% +3.6867% +3.8658%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  3 (3.00%) high mild

     Running /home/vertexclique/projects/arrow/rust/target/release/deps/arithmetic_kernels-ce51190c627182f2
add 2^20                time:   [971.06 us 972.57 us 974.19 us]                     
                        change: [-58.352% -58.310% -58.264%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) low mild
  1 (1.00%) high mild

subtract 2^20           time:   [970.69 us 972.03 us 973.30 us]                          
                        change: [-58.416% -58.371% -58.327%] (p = 0.00 < 0.05)
                        Performance has improved.

multiply 2^20           time:   [977.12 us 978.10 us 979.03 us]                          
                        change: [-58.449% -58.374% -58.308%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking divide 2^20: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 8.7s, enable flat sampling, or reduce sample count to 50.
divide 2^20             time:   [1.7273 ms 1.7275 ms 1.7278 ms]                         
                        change: [-53.618% -53.550% -53.507%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  12 (12.00%) high mild
  2 (2.00%) high severe

limit 2^20, 512         time:   [114.62 ns 114.65 ns 114.69 ns]                            
                        change: [+4.5154% +4.6662% +4.7672%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low severe
  4 (4.00%) high mild
  1 (1.00%) high severe

add_nulls_2^20          time:   [980.77 us 982.49 us 984.18 us]                           
                        change: [-58.295% -58.224% -58.159%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking divide_nulls_2^20: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 8.7s, enable flat sampling, or reduce sample count to 50.
divide_nulls_2^20       time:   [1.7293 ms 1.7295 ms 1.7297 ms]                               
                        change: [-54.156% -54.132% -54.101%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe


     Running /home/vertexclique/projects/arrow/rust/target/release/deps/array_from_vec-45cab0a3ecf39249
array_from_vec 128      time:   [484.23 ns 484.68 ns 485.12 ns]                               
                        change: [+2.1162% +2.2409% +2.3687%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

array_from_vec 256      time:   [712.61 ns 713.50 ns 714.36 ns]                                
                        change: [+2.1834% +2.3743% +2.5626%] (p = 0.00 < 0.05)
                        Performance has regressed.

array_from_vec 512      time:   [1.1879 us 1.1892 us 1.1906 us]                                
                        change: [-4.2637% -4.1494% -4.0486%] (p = 0.00 < 0.05)
                        Performance has improved.

array_string_from_vec 128                                                                             
                        time:   [2.2701 us 2.2709 us 2.2716 us]
                        change: [+6.6458% +6.7030% +6.7594%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild

array_string_from_vec 256                                                                             
                        time:   [3.6563 us 3.6570 us 3.6577 us]
                        change: [+5.7356% +5.7737% +5.8114%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

array_string_from_vec 512                                                                             
                        time:   [6.1225 us 6.1246 us 6.1268 us]
                        change: [+5.6870% +5.7640% +5.8419%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe

struct_array_from_vec 128                                                                            
                        time:   [117.76 us 118.15 us 118.53 us]
                        change: [+3455.9% +3472.4% +3491.6%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe

struct_array_from_vec 256                                                                            
                        time:   [105.76 us 106.43 us 107.21 us]
                        change: [+2015.8% +2029.9% +2045.7%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe

struct_array_from_vec 512                                                                             
                        time:   [9.0257 us 9.0293 us 9.0328 us]
                        change: [+23.421% +23.496% +23.570%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe

struct_array_from_vec 1024                                                                             
                        time:   [27.670 us 27.759 us 27.859 us]
                        change: [+134.39% +135.24% +136.13%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

     Running /home/vertexclique/projects/arrow/rust/target/release/deps/array_slice-49615c84b19508ee
array_slice 128         time:   [163.62 ns 163.71 ns 163.80 ns]                            
                        change: [+28.932% +28.988% +29.037%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 17 outliers among 100 measurements (17.00%)
  5 (5.00%) low severe
  4 (4.00%) low mild
  6 (6.00%) high mild
  2 (2.00%) high severe

array_slice 512         time:   [265.55 ns 265.63 ns 265.71 ns]                            
                        change: [+98.289% +98.369% +98.439%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

array_slice 2048        time:   [649.40 ns 650.84 ns 652.87 ns]                              
                        change: [+319.75% +320.19% +320.76%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  5 (5.00%) high severe

     Running /home/vertexclique/projects/arrow/rust/target/release/deps/boolean_kernels-6fc1ad233dfa3072
and                     time:   [34.437 us 34.447 us 34.457 us]                 
                        change: [+39.672% +39.729% +39.788%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild

or                      time:   [34.540 us 34.550 us 34.560 us]                
                        change: [+39.274% +39.429% +39.552%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

not                     time:   [17.377 us 17.386 us 17.393 us]                 
                        change: [+38.007% +38.289% +38.681%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

     Running /home/vertexclique/projects/arrow/rust/target/release/deps/buffer_bit_ops-7f2013fd9e28372f
buffer_bit_ops and      time:   [28.631 us 28.749 us 28.893 us]                                
                        change: [+4064.1% +4101.4% +4148.8%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe

     Running /home/vertexclique/projects/arrow/rust/target/release/deps/builder-b9aacab6b54f093d
Benchmarking bench_primitive: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.7s, enable flat sampling, or reduce sample count to 50.
bench_primitive         time:   [1.5190 ms 1.5193 ms 1.5195 ms]                             
                        thrpt:  [2.5707 GiB/s 2.5711 GiB/s 2.5715 GiB/s]
                 change:
                        time:   [+20.305% +20.339% +20.372%] (p = 0.00 < 0.05)
                        thrpt:  [-16.924% -16.902% -16.878%]
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low severe
  1 (1.00%) high mild
  2 (2.00%) high severe

bench_bool              time:   [3.7009 ms 3.7045 ms 3.7080 ms]                        
                        thrpt:  [134.84 MiB/s 134.97 MiB/s 135.10 MiB/s]
                 change:
                        time:   [+17.329% +17.443% +17.585%] (p = 0.00 < 0.05)
                        thrpt:  [-14.955% -14.853% -14.770%]
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild

     Running /home/vertexclique/projects/arrow/rust/target/release/deps/cast_kernels-93bbf3677ef76b84
cast int32 to int32 512 time:   [29.420 ns 29.421 ns 29.423 ns]                                     
                        change: [-0.5265% -0.4906% -0.4550%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  3 (3.00%) high mild
  5 (5.00%) high severe

cast int32 to uint32 512                                                                             
                        time:   [12.053 us 12.055 us 12.057 us]
                        change: [+126.57% +126.69% +126.87%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

cast int32 to float32 512                                                                             
                        time:   [12.306 us 12.309 us 12.312 us]
                        change: [+115.72% +115.86% +115.97%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

cast int32 to float64 512                                                                             
                        time:   [12.389 us 12.391 us 12.393 us]
                        change: [+133.82% +133.93% +134.04%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

cast int32 to int64 512 time:   [12.402 us 12.405 us 12.407 us]                                     
                        change: [+126.67% +126.81% +126.92%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

cast float32 to int32 512                                                                             
                        time:   [12.765 us 12.767 us 12.768 us]
                        change: [+114.01% +114.11% +114.20%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild

cast float64 to float32 512                                                                             
                        time:   [12.522 us 12.524 us 12.526 us]
                        change: [+133.22% +133.35% +133.47%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

cast float64 to uint64 512                                                                             
                        time:   [13.187 us 13.191 us 13.194 us]
                        change: [+100.11% +100.20% +100.30%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

cast int64 to int32 512 time:   [11.755 us 11.758 us 11.761 us]                                     
                        change: [+122.67% +122.77% +122.87%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

cast date64 to date32 512                                                                             
                        time:   [23.245 us 23.248 us 23.252 us]
                        change: [+37.888% +37.929% +37.969%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

cast date32 to date64 512                                                                             
                        time:   [22.948 us 22.951 us 22.954 us]
                        change: [+35.232% +35.391% +35.493%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

cast time32s to time32ms 512                                                                             
                        time:   [1.1747 us 1.1750 us 1.1753 us]
                        change: [-34.441% -34.422% -34.402%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe

cast time32s to time64us 512                                                                             
                        time:   [13.960 us 13.961 us 13.962 us]
                        change: [+79.328% +79.399% +79.468%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

cast time64ns to time32s 512                                                                             
                        time:   [25.761 us 25.765 us 25.768 us]
                        change: [+19.682% +19.723% +19.766%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe

cast timestamp_ns to timestamp_s 512                                                                             
                        time:   [29.816 ns 29.825 ns 29.834 ns]
                        change: [+0.3917% +0.4310% +0.4744%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  6 (6.00%) high mild
  2 (2.00%) high severe

cast timestamp_ms to timestamp_ns 512                                                                             
                        time:   [1.5295 us 1.5299 us 1.5304 us]
                        change: [-17.402% -17.337% -17.272%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

cast timestamp_ms to i64 512                                                                            
                        time:   [176.67 ns 176.84 ns 177.19 ns]
                        change: [-1.7586% -1.6436% -1.4924%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

     Running /home/vertexclique/projects/arrow/rust/target/release/deps/comparison_kernels-f8939ec12975f45e
eq Float32              time:   [36.413 us 36.429 us 36.447 us]                        
                        change: [-95.357% -95.353% -95.349%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

eq scalar Float32       time:   [33.540 us 33.551 us 33.562 us]                               
…