@nevi-me 2020-11-15
@vertexclique did you work on this on top of #8663 from @jhorstmann (or at least both PRs remove the popcnt table)?
The changes here look reasonable to me, so I can review after a rebase, as I've just merged #8663