@vertexclique 2020-11-15
This is using safe bitvec interface to manage bits. I worked on top of my existing PR. Since that didn't change all code to use safe operations, now this does. I squashed all my work into a single commit and closing PR: https://github.com/apache/arrow/pull/8598

Closed that with comment:  https://github.com/apache/arrow/pull/8598#issuecomment-727294219