@jorgecarleitao 2020-11-18
Super cool. Great work @vertexclique ! 💯  I am all in in removing that `unsafe` code. 👍 

Just curious, do you know why it has a 120% hit in performance on a filtering op?

```
filter u8 high selectivity                                                                             
                        time:   [18.523 us 18.529 us 18.535 us]
                        change: [+122.09% +122.22% +122.33%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
```