@vertexclique 2020-11-18
Finished, I have addressed all the comments. @jorgecarleitao I saw that it is because of the unaligned bit checks and boundary checks in byte slices. Yet another addition comes from `as_buffer` method. Since we are allocating a buffer for it now to make it well aligned this is creating an overhead. I have tried to overcome that problem by writing unowned buffer but one test failed. I am also unsure that test is actually testing correctly or not.