@alamb 2020-11-18
I ran the filtering selectivity benchmark on my laptop again and this time it shows a ~400% slowdown... So I wonder if that unsafe code is actually giving us a performance gain?

results on master @ 0e2681fe52ea48cade10171a018554da05f95285
```
cargo bench -- "filter u8 high selectivity"

filter u8 high selectivity
                        time:   [3.5905 us 3.6379 us 3.6942 us]
                        time:   [3.3929 us 3.4179 us 3.4457 us]
                        time:   [3.3876 us 3.4167 us 3.4470 us]
                        time:   [3.4520 us 3.4747 us 3.4981 us]
                        time:   [3.4288 us 3.4538 us 3.4820 us]

```

Then on this PR at @457d8c7303864673c25bcb4e95e002b9585e9c39

```
filter u8 high selectivity
                        time:   [16.951 us 17.164 us 17.395 us]
                        time:   [16.649 us 16.886 us 17.136 us]
                        time:   [17.175 us 17.751 us 18.520 us]
                        time:   [16.028 us 16.109 us 16.196 us]
                        time:   [16.224 us 16.341 us 16.456 us]
```

Which does seem like a substantial and significant performance regression.

I double checked that I could reproduce the faster master numbers and I can