@vertexclique 2020-11-18
Weird, on my machine when I pushed the initial implementation of this PR I got the numbers above. Seems like it is regressed for me too.