@vertexclique 2020-11-19
So any idea how to proceed with this? Based on that I will close all related PRs. Though, seems like https://github.com/apache/arrow/pull/8688 can use these changes to build slice alignment on top.