@nevi-me 2020-11-19
If this is not stalling any other PRs, you can leave it open. I've started looking into what's causing the regression, but I'll only finish on the weekend (I'm mostly working on the Parquet writer with my evening time).

I'm also seeing big regressions of 400%+