@vertexclique 2020-11-24
@nevi-me @alamb @jorgecarleitao 
The good news, I have found a solution for the performance related considerations. I have experimented on the `sum` and my roofline analysis brought some good results. Also, criterion benches are here:

Before:

```
sum 2^20                time:   [900.13 us 902.01 us 904.02 us]                     
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  4 (4.00%) high severe

sum nulls 2^20          time:   [2.5859 ms 2.5909 ms 2.5967 ms]                            
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe
```

After:

```
sum 2^20                time:   [236.61 us 238.02 us 239.58 us]                     
                        change: [-73.888% -73.699% -73.493%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  5 (5.00%) high severe

sum nulls 2^20          time:   [549.14 us 551.39 us 554.07 us]                           
                        change: [-78.784% -78.671% -78.548%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  7 (7.00%) high mild
  5 (5.00%) high severe
```

Since it is a time-consuming task, I am not going to perform a full rewrite until we agree that this performance improvement is enough.