@nevi-me 2020-11-24
Thanks @vertexclique, that looks great. I'm here with the approach