@jhorstmann 2020-11-25
> Introducing thread parallelism at a compute kernel level

Conceptually this is not a small change. Personally I think parallelizing on the datafusion level and keeping kernels single-threaded is the better model.

The benchmark is now also testing a much larger array (2^20 elements) than what is usually used as a chunk size, so in reality the speedup due to parallelism would be much smaller.

I'm totally fine with a small slowdown if that leads to cleaner and safer code.